refactor : 사용자 정의 타입 제거
