+++ README.md
@@ -16,27 +16,24 @@
 
 ### STEP 1
 #### 고민했던 점
-1. `struct GameService`
--- 현 상황에서 GameService가 상속관계를 갖지 않기 때문에 값타입으로 존재하는 `struct`를 사용하였습니다.
-
-2. 컴퓨터가 `게임 시작 시 생성한 임의의 수`를 의미하는 `var answer` 네이밍
+1. 컴퓨터가 `게임 시작 시 생성한 임의의 수`를 의미하는 `var answer` 네이밍
 -- "정답"이 있는 게임이라고 생각해서 `answer`라고 작명하였는데, 추후에 플레이어의 입력을 받게 될 경우 플레이어가 입력한 답과 비교할 때 네이밍이 헷갈릴 수도 있다는 의견이 있었습니다.
 -- 더 나은 대안을 찾지 못해 `answer`라고 작성하였는데, 리뷰어님의 의견이 궁금합니다.
 
-3. `var isPlayerWin`
+2. `var isPlayerWin`
 -- 기존에 작성하였던 게임 결과를 판단하고 출력하는 함수인 `func printGameResult`를 이용해 3 Strike시 `while` 문을 탈출할 수 있도록 `Bool`값을 반환하도록 할지 고민했습니다.
 -- 하지만 해당 함수에서 `Bool`값을 반환한 후 while문에서 이 반환값을 사용할 경우, 함수를 별도로 분리한 의의가 퇴색되는 면이 있다고 생각하여 `var isPlayerWin`이라는 새로운 연산 프로퍼티를 사용하기로 했습니다.
 -- `Bool`의 값이기 때문에 단순히 `is`를 붙여서 `플레이어가 이겼는가?`라는 의미의 변수 네이밍을 사용하였습니다.
 
-4. `func decideBallStrike`
+3. `func decideBallStrike`
 -- 볼/스트라이크 여부를 판정하는 함수 `func validationNumber`와 각 숫자를 비교하는 함수 `func compareNumbers`의 역할을 통합해서 하나의 함수로 변경하였습니다. 그 이유는 기존에 있던 `func validationNumer`의 매개변수로 존재했던 index값이 `func compareNumbers`에서 가공을 하지 않는 값이기 때문에 통합을 통해서 불필요한 매개변수를 없애고자 하였습니다.
 -- 함수 네이밍 시 `결정하다.`라는 의미의 단어를 사용하였으며, ball과 strike를 모두 판정하기 때문에 변수명에 모두 포함되도록 구분하였습니다.
 
-5. `func generateDescription`
+4. `func generateDescription`
 -- 각 회차마다 선택된 임의의 수 3개를 출력할 때, `[Int]`를 그대로 출력할 경우 `[1, 2, 3]`과 같이 출력되기 때문에 `1 2 3`과 같이 출력할 수 있도록 `[Int]`를 변환해주는 메서드를 작성하였습니다.
 -- `toString()`이라는 이름을 지양하고자, Array의 값을 표현해주는 결과물을 생성한다는 의미에서 `generateDescription`으로 이름지었습니다.
 
-6. 판정 결과 출력(볼/스트라이크 개수), 남은 기회 출력을 별도의 함수로 작성할 것인가 / `startGame()` 함수 내에 포함하여 작성할 것인가?
+5. 판정 결과 출력(볼/스트라이크 개수), 남은 기회 출력을 별도의 함수로 작성할 것인가 / `startGame()` 함수 내에 포함하여 작성할 것인가?
 -- 처음에는 판정 결과 출력 함수를 별도로 작성하고, 해당 함수 내에서 플레이어 승리 여부를 판단하도록 작성했습니다.
 -- 하지만 `var isPlayerWin`을 통해 플레이어 승리 여부를 판단하도록 구조를 변경하면서, 판정 결과 출력 함수에서는 플레이어 승리 여부는 판단하지 않고 결과 출력만 담당하게 되어 **출력만을 위한 함수를 작성하는 것이 타당한가?** 에 대해 이야기를 나눈 결과 현재와 같이 출력하는 함수는 별도로 작성하지 않기로 했습니다.
 
@@ -46,3 +43,4 @@
 3. **Array Extension 내에서 함수를 사용하는 것이 아니라 변수를 사용하여 표현하는 것**에 대한 리뷰어님의 의견이 궁금합니다.
 
 ![](https://i.imgur.com/LZog4dt.png)
+
